# Add a fast path for ASCII strings