@headius 2025-01-08
Did this get applied to the Java parser anywhere? I can take it if not but I don't want to dive in and waste my time if it's already there.