@headius 2025-01-09
Work is proceeding on this and other optimizations for the JRuby extension in https://github.com/ruby/json/pull/725.